Shifted three functions to common (#1101)
